Add files via upload
I don't know how to edit the code in already uploaded file so I uploaded a new one. I'm a newbie.
Anyways I'm open to constructive criticism in the code. If the code is not working properly, kindly let me know the exact scenario so I can fix it.
Thanks.